requirements.txt
requests~=2.31.0
gitpython
argparse~=1.4.0
boto3
botocore